pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nanobanana"
version = "0.0.0.dev0"
description = "CLI tool for generating and editing images using Google's Gemini API or OpenRouter"
readme = "README.md"
license = "MIT"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
nanobanana = "nanobanana.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/nanobanana"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = ["pytest>=8.0"]
